+++ jest.preset.js
@@ -16,9 +16,13 @@ module.exports = {
     "^@sentrei/organisms": "<rootDir>/../../components/organisms/src/index.ts",
     "^@sentrei/roots": "<rootDir>/../../components/roots/src/index.ts",
     "^@sentrei/screen": "<rootDir>/../../components/screens/src/index.ts",
-    "^@sentrei/app/(.+)$": "<rootDir>/../../apps/app/src/$1",
-    "^@sentrei/demo/(.+)$": "<rootDir>/../../apps/demo/src/$1",
-    "^@sentrei/home/(.+)$": "<rootDir>/../../apps/home/src/$1",
+    "^@sentrei/app-sentrei-com/(.+)$":
+      "<rootDir>/../../apps/app-sentrei-com/src/$1",
+    "^@sentrei/blog-sentrei-com/(.+)$":
+      "<rootDir>/../../apps/blog-sentrei-com/src/$1",
+    "^@sentrei/og-sentrei-com/(.+)$":
+      "<rootDir>/../../apps/og-sentrei-com/src/$1",
+    "^@sentrei/www-mosh-lol/(.+)$": "<rootDir>/../../apps/www-mosh-lol/src/$1",
   },
   transform: {
     "^.+\\.[tj]sx?$": "ts-jest",